Add programmatic command and option completion
This is written in bash with the commands and options hardcoded so
that the completions are generated very quickly. I did try using the
completion generation provided by yargs, but it took too long (around
1 second), which is very frustrating as a user. Completions should
improve the user experience, not detract from it.

This requires a bash script to be appended to ~/.bashrc to be
persisted. This is set up automatically when layer is installed. The
bash script to be appended is based on the one provided by yargs.

Why bash?
==============

The speed issue is primarily due to layer being executed as
coffeescript, which is not quick on my machine. For comparison, here
are the times to print 'test' for different languages on my machine:

    ------------------------------------------------------------
    || Command                              || Time (real)    ||
    ||--------------------------------------||----------------||
    || echo "console.log 'test'" | coffee   || ~0.500 seconds ||
    || echo "console.log('test')" | node    || ~0.270 seconds ||
    || echo "puts 'test'" | ruby            || ~0.085 seconds ||
    || echo "print 'test'" | python         || ~0.030 seconds ||
    || echo "echo test" | bash              || ~0.005 seconds ||
    ------------------------------------------------------------

Bash definitely seems like the fastes option and completions should be
as fast as possible. As the completions need to be passed to bash by
writing to the console, this seems like a fair enough test.