Store user-writable files in the user's home directory
Previously, some of layer's functionality depended on writing to files
that were installed by npm. By default, these files are globally
installed in a directory owned by root (/usr/local), meaning that root
permisions are necessary to write to these files. However, requiring
users to 'sudo' everything that requires writing seems wrong, so these
config files need to be owned by the user. Due to permission problems,
I was not able to copy the config files to the correct directory
during installation, so instead the first thing layer does is copy
them to the correct place if they do not exist. This also adds a bit
of resilience to the config directory being removed.