Only use template-data and template-jade options in 'new'
The defaults of these options were overriding the arguments specified
when just a template name was provided, preventing the template option
from behaving as expected in many commands. The ability to specify
overrides isn't really necessary, so has been removed.

These options remain in the 'new' command as there needs to be some
way to specify where a template's components are located.